+++ renderers/lit/src/v0_9/catalogs/basic/components/Text.ts
@@ -138,7 +138,7 @@ export class A2uiBasicTextElement extends BasicCatalogA2uiLitElement<typeof Text
     // There's not a good way to handle the caption variant in markdown, so we
     // tag it with a class so it can be tweaked via CSS.
     if (props.variant === 'caption') {
-      return html`<span class="a2ui-caption">${renderedMarkdown}</span>`;
+      return html`<span class="a2ui-caption"><em>${renderedMarkdown}</em></span>`;
     }
     return html`${renderedMarkdown}`;
   }